refactor(invite): fix invited client login fixed, remove email templates duplications.
